Add deserialize_and_run functionality and update submission scripts
- Implemented `deserialize_and_run` function in CLI to handle task deserialization and execution.
- Updated `sge_submit.sh` and `slurm_submit.sh` scripts to use the new deserialization command.

Note: This update breaks cross-compatibility with ribbon versions below v0.3.0 and Tasks versions below v0.1.4. This is because previously, on deserialization the container would run the HOST's deserialization script in the CONTAINER's environment. The new format for passing in configs requires an update to how we handle deserialization, so running the new deserialization code in an old environment does not work.